backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["diyanovamir <amirdianov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = ">=4.2,<5.0"
psycopg2 = ">=2.9.6,<2.10.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
requests = "^2.28.2"
django-cors-headers = "^3.14.0"
drf-yasg = "^1.21.5"
python-dotenv = "^1.0.0"
celery = "^5.2.7"
redis = "^4.5.5"
django-celery-beat = "^2.5.0"
channels = "^4.0.0"
channels-redis = "^4.1.0"
daphne = "^4.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pre-commit = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
